# speed up to-json and add to-json-file