to-json-file will help with large data because it's not keeping tons of temporary Strs around, nor does it concat.

Benchmark by commit:
b456d145f85bd90b0284d7400532ab6d4d7934ca 15.877100s
4ed3b04430e9dc26a7deb2bf0fc66b63c75d4d7f 61.9206466s
a83c4fd7349f0b566bd383a4a0aaac858c8d7b76 61.6610808s
43593ffb24db6d603616f4966715030e8937a4b3 67.9428612s
cbb3536cccbc7c144039fca399cd969b3e7355de 88.3700205s
4ef1e02bc467280567d7a366db68a091bea627c6 97.52776943s