@lizmat 2021-09-10
I think this is superseded by all of the optimizations that have been done since then.  So I assume this can be closed?